validator done

+++ src/by/academy/lesson21/notebook/logic/impl/NotebookLogicImpl.java
@@ -15,54 +15,40 @@ import java.util.List;
 public class NotebookLogicImpl implements NotebookLogic {
     private final DaoProvider provider = DaoProvider.getInstance();
     private final NoteBookDao dao = provider.getNoteBookDao();
-    private boolean isDataRead = false;
+
+    public NotebookLogicImpl() {
+    }
 
     @Override
-    public String add(String header, String text, LocalDateTime creationDate) throws NotebookLogicException {
+    public String add(String header, String text, LocalDateTime creationDate) {
         Note note;
 
-        try {
-            note = new Note(header, text, creationDate);
-            dao.add(note);
-            //dao.save();
-        } catch (NoteBookException e) {
-            throw new NotebookLogicException(e);
-        }
+        note = new Note(header, text, creationDate);
+        dao.add(note);
 
-        //return "Не удалось добавить запись.";
         return "Запись успешно добавлена.";
     }
 
     @Override
-    public String updateByIndex(String number, String header, String text) throws NotebookLogicException {
-        int index = 0;
+    public String updateByIndex(int index, String header, String text) {
         Note note;
 
-        try {
-            index = Integer.parseInt(number) - 1;
-            note = dao.allNotes().get(index);
-        } catch (NumberFormatException e) {
-            throw new NotebookLogicException("Указан неправильный номер.", e);
-        } catch (IndexOutOfBoundsException e) {
-            throw new NotebookLogicException("Неправильно указан номер.", e);
+        if (index + 1 > dao.allNotes().size()) {
+            return "Неправильно указан номер.";
         }
+        note = dao.allNotes().get(index);
         note.setHeader(header);
         note.setText(text);
-        //dao.save();
+
         return "Запись успешно обновлена";
     }
 
     @Override
-    public String delete(String number) throws NotebookLogicException {
-        int index = 0;
-        try {
-            index = Integer.parseInt(number) - 1;
-            dao.delete(index);
-        } catch (NumberFormatException e) {
-            throw new NotebookLogicException("Неправильно указан номер.", e);
-        } catch (NoteBookException e) {
-            throw new NotebookLogicException(e);
+    public String delete(int index) {
+        if (index + 1 > dao.allNotes().size()) {
+            return "Неправильно указан номер.";
         }
+            dao.delete(index);
         return "Запись успешно удалена";
     }
 
@@ -114,25 +100,7 @@ public class NotebookLogicImpl implements NotebookLogic {
     }
 
     @Override
-    public String read() throws NotebookLogicException {
-        if (!isDataRead) {
-            try {
-                dao.read();
-                isDataRead = true;
-            } catch (NoteBookException e) {
-                throw new NotebookLogicException(e);
-            }
-        }
-
-        return noteToString(dao.allNotes());
-    }
-
-    @Override
     public String save() throws NotebookLogicException {
-        if (!isDataRead){
-            return "Сохранять нечего.";
-        }
-
         try {
             dao.save();
         } catch (NoteBookException e) {
